Update managed-operating-system-frequently-asked-questions.md
feedback from the field (please have product confirm) is that managed servers are joined to a domain on build.  in order for a client to use their own domain they must deploy a managed active directory.  the managed teams have told field sales they don't support customers deploying an non-managed AD and then joining managed servers to it since they'd lose the ability to manage the customer VMs and don't want to deal with trusts to non-managed AD structures.