Cloudera KB Article Updates
Updated the getting started with Cloudera KB article with up to date
screenshots, current information on add-on components and added
additional Q & A's. Added links to the VPN KB article and the Cloudera
PDF in the attachments folder.
Updated the PDF link in the Cloudera Manager Guide KB article.
Created a Cloudera folder in the images folder and added the six
screenshots.

Thanks, Tim Baumgartner
tim.baumgartner@centurylink.com