Adding region information, small change to wording in intro
I am adding the datacenters that make up a region in the service point
KB article.  Because of this, I am renaming the KB article.  I have
updated the relevant links to that KB article to point to the correct
file.  I also made a small change to the intro page.