update roles matrix per may 24th release
per may 24th release, included Anti-Affinity as well since this was missed
Minor change to the Server Operator role in Control. The Server Operator role now has the ability to CREATE, EDIT and DELETE Vertical Autoscale, Horizontal Autoscale, and Alert Policies in Control (previously, it was VIEW only). This is a minor change which has no impact on cost or other features within the Control Portal.