Update how-to-configure-client-vpn.md
Added the following to Q & A to clarify the requirement for OpenVPN server, as well as to inform the customer that the server is already built for them. Please check to be sure the formatting is correct.

Q: Do I need to install an OpenVPN server?

A: No, every customer account receives a dedeicated OpenVPN server by default. Please note that it cannot be removed or disabled due to platform dependencies.