Update building-coreos-server-cluster-on-the-centurylink-cloud.md
Recommend adding the following note under "limitations"

"CoreOS is a "read only" OS, so normal methods of expanding filesystems will not be successful. To increase space on this OS, you will need to add space to the 0:0 drive via control, and then reboot the machine. Adding space to the 0:1 partition, or adding additional drives is not possible at this time".