Update how-do-i-report-a-support-issue.md
Hi. I'm trying to fix the "chat system requirements" link as is it broken on the current page. Clicking the link does not take you anywhere. I changed the indentation of the section by adding another "#". The link at the top is indented 2x and the section appears to be indented only once. I am not sure if this solution will work. Can you please verify? I will post in the slack digital channel on Monday for confirmation. Thank you.