Update network-exchange-clc-colocation-endpoint-guide.md
Removed the following text:  • Up to 10Gb/s of shared bandwidth to and from the colocation site. Network performance is offered as “best effort”.
Added/Modified following:  •	The End User must initiate a request for two instances of "HAE" (Hosting Access Extension) - a cross connect - between Network Exchange and the End User’s equipment following the provisioning of Network Exchange. Network Exchange will provide the equipment location and ports on the Network Exchange switch(es) to be connected to, so the connections can be completed. 
•	CSCO-GLC-T  1000BASE-T SFP Copper SFP, Up to 100M (328 ft.) on Cat 5 unshielded twisted-pair copper cabling
 ...... and type (SMF or Cat 5)
See the Network Exchange Availability Matrix and Configuration Guide for available endpoints per metro area.