Updated KB with new flow
We've made changes to the login flow and this KB update is reflecting that.  The first change is to allow users to see where their email is during the registration process.  The second change is instead of directing unlinked accounts to a login page that looks like a typical login page, the user will now be redirected to the account linking page.