Updated KB with new flow (#2375)
* Updated KB with new flow

We've made changes to the login flow and this KB update is reflecting that.  The first change is to allow users to see where their email is during the registration process.  The second change is instead of directing unlinked accounts to a login page that looks like a typical login page, the user will now be redirected to the account linking page.

* Adding a screenshot

Added a screenshot of the new account linking page.

* Updating screenshots

The flow for the master account creation has slightly changed.  These screenshots reflect the new flow.