Initial commit and updating KBs for SMS (#2441)
* Updating for SMS functionality

Adding a section for SMS errors.

* Initial commit of new KB

This is the initial commit for the new KB.  This KB walks a user through the process of setting up their account to be able to reset their password using a token delivered via SMS.

* Initial commit for screenshots for new KB

These are the screenshots used in the new KB, Enhanced Security - Enabling Ability to Reset Account via SMS

* Testing fix to images

* Fixing the rest of the images

The test worked, fixing the rest.

* Changing title of the file

* Updating screenshot

The account creation process now allows you to setup a PIN.  Updating screenshot to reflect that.

* Updating KB

You now get a PIN assigned and can change it during the registration.  The steps have been updated to reflect that.