"Close Account" link in control is not reliable
"Close Account" link in control is not reliable. 
It creates a poor customer experience to think that the user has done everything necessary to request account closure when in fact no ticket was generated from the click.

So updating the Public KB to guide the customer to manually send an email to help@ctl.io or submit a separate ticket to request the account to be closed.