Edited the **Resolution of Full Failure** section for clarity
Removed the phrasing "If there's no data that needs to be recovered" because it implies that we can assist with recovering data.  
Data backup and recovery is the customer's responsibility and is outside the scope of support for the Edge Bare Metal product. In the case of critical hardware failure, we recommend reprovisioning new hardware as we will 'not' attempt to recover the current bare metal server.